Factored out analyser macros and helpers to separate files.
